Refactored mod integration code a little more.
