+++ src/main/java/growthcraft/core/GrowthcraftCore.java
@@ -3,8 +3,8 @@ package growthcraft.core;
 import growthcraft.core.shared.GrowthcraftCoreApis;
 import growthcraft.core.shared.Reference;
 import growthcraft.core.shared.compat.Compat;
-import growthcraft.core.shared.compat.InitForestry;
-import growthcraft.core.shared.compat.InitRustic;
+import growthcraft.core.shared.compat.forestry.InitForestry;
+import growthcraft.core.shared.compat.rustic.InitRustic;
 import growthcraft.core.common.CommonProxy;
 import growthcraft.core.common.Init;
 import growthcraft.core.common.creativetabs.TabGrowthcraft;
