+++ src/main/java/growthcraft/core/shared/compat/rustic/RusticModItems.java
@@ -1,6 +1,5 @@
 package growthcraft.core.shared.compat.rustic;
 
-import growthcraft.core.shared.definition.FluidDefinition;
 import growthcraft.core.shared.definition.ItemDefinition;
 
 public class RusticModItems {
